requirements.txt
aiogram==2.11.2
aiohttp==3.7.4.post0
APScheduler==3.7.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==3.0.4
Django==3.2.5
django-cors-headers==3.8.0
djangorestframework==3.12.4
environs==9.3.2
idna==2.10
loguru==0.5.3
marshmallow==3.13.0
multidict==5.1.0
psycopg2==2.9.1
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.7
yarl==1.6.3
